scripts/requirements.txt
openai>=1.30.0
edge-tts>=6.1.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
